feat: renamed forum database and fixed multiple api (#230)
* feat: updated api get schema and renamed database

* fix: fix lambda function

* fix: fixed lambda get single thread and others
