+++ src/lambda/delete-thread/index.py
@@ -1,0 +1,2 @@
+def handler(event, context):
+    pass
